+++ container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/enclave_object_attributes_provider.go
@@ -22,7 +22,7 @@ import (
 
 const (
 	artifactExpansionObjectTimestampFormat = "2006-01-02T15.04.05.000"
-	userServiceServiceSuffix = "kurtosis-user-service-service"
+	userServiceSuffix = "user-service"
 
 )
 
@@ -30,7 +30,7 @@ type KubernetesEnclaveObjectAttributesProvider interface {
 	ForEnclaveNamespace(isPartitioningEnabled bool) (KubernetesObjectAttributes, error)
 	ForEnclaveDataVolume() (KubernetesObjectAttributes, error)
 	ForApiContainer() (KubernetesApiContainerObjectAttributesProvider, error)
-	ForUserServiceService(id string, guid string) (KubernetesObjectAttributes, error)
+	ForUserServiceService(id user_service_registration.ServiceID, guid service.ServiceGUID) (KubernetesObjectAttributes, error)
 }
 
 // Private so it can't be instantiated
@@ -160,34 +160,34 @@ func (provider *kubernetesEnclaveObjectAttributesProviderImpl) ForFilesArtifactE
 }
 
 func (provider *kubernetesEnclaveObjectAttributesProviderImpl) ForUserServiceService (
-	id string,
-	guid string,
+	serviceID user_service_registration.ServiceID,
+	serviceGUID service.ServiceGUID,
 ) (
 	KubernetesObjectAttributes,
 	error,
 ) {
-	name, err := provider.getNameForEnclaveObject([]string{userServiceServiceSuffix})
+	name, err := provider.getNameForEnclaveObject([]string{userServiceSuffix})
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "Failed to get name for user service registration.")
+		return nil, stacktrace.Propagate(err, "Failed to get name for user service service.")
 	}
 
-	labels, err := provider.getLabelsForEnclaveObjectWithIDAndGUID(id, guid)
+	labels, err := provider.getLabelsForEnclaveObjectWithIDAndGUID(string(serviceID), string(serviceGUID))
 	if err != nil {
 		return nil, stacktrace.Propagate(
 			err,
-			"Failed to get labels for user service registration with ID '%s' and GUID '%s'.",
-			id,
-			guid,
+			"Failed to get labels for user service service with ID '%s' and GUID '%s'.",
+			string(serviceID),
+			string(serviceGUID),
 		)
 	}
-	labels[label_key_consts.KurtosisUserServiceServiceTypeLabel] = label_value_consts.UserServiceServiceKurtosisResourceTypeLabelValue
+	labels[label_key_consts.KurtosisResourceTypeLabelKey] = label_value_consts.UserServiceKurtosisResourceTypeLabelValue
 
-	//No userServiceRegistration annotations.
+	//No userServiceService annotations.
 	annotations := map[*kubernetes_annotation_key.KubernetesAnnotationKey]*kubernetes_annotation_value.KubernetesAnnotationValue{}
 
 	objectAttributes, err := newKubernetesObjectAttributesImpl(name, labels, annotations)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "Failed to create user service registration object attributes.")
+		return nil, stacktrace.Propagate(err, "Failed to create user service service object attributes.")
 	}
 
 	return objectAttributes, nil